requirements.txt
Flask
numpy
google-generativeai
scikit-learn
